backend/requirements.txt
fastapi[all]
sqlalchemy[asyncio]
psycopg2-binary
alembic
openai
ollama
python-dotenv
pytest
httpx